Names to constants-file, little changes
    - typos
    - remove "boldText" labels in translation files

Ticket: #307
